+++ src/api/auth.ts
@@ -1,8 +1,8 @@
-import type { AuthUser } from '@/store/useAuthStore'
+import type { User } from '@/types'
 
 import { api } from './api'
 
-export const getMe = async (): Promise<AuthUser> => {
+export const getMe = async (): Promise<User> => {
   const res = await api.get('/accounts/me')
   return res.data
 }
