+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export { ROUTES_PATHS } from './router'
