+++ src/types/index.ts
@@ -4,4 +4,6 @@ export * from './api-response/chat'
 export * from './api-response/chatbot'
 export * from './api-response/detail'
 export * from './api-response/qna'
+export * from './refresh-qna'
 export * from './toast'
+export * from './user'
